--- a/frappe/www/about.py
+++ b/frappe/www/about.py
@@ -7,6 +7,6 @@ sitemap = 1
 
 
 def get_context(context):
-	context.doc = frappe.get_doc("About Us Settings", "About Us Settings")
+	context.doc = frappe.get_cached_doc("About Us Settings")
 
 	return context