--- a/frappe/www/app.py
+++ b/frappe/www/app.py
@@ -32,8 +32,6 @@ def get_context(context):
 
 	frappe.db.commit()
 
-	desk_theme = frappe.db.get_value("User", frappe.session.user, "desk_theme")
-
 	boot_json = frappe.as_json(boot)
 
 	# remove script tags from boot
@@ -52,7 +50,7 @@ def get_context(context):
 			"lang": frappe.local.lang,
 			"sounds": hooks["sounds"],
 			"boot": boot if context.get("for_mobile") else boot_json,
-			"desk_theme": desk_theme or "Light",
+			"desk_theme": boot.get("desk_theme") or "Light",
 			"csrf_token": csrf_token,
 			"google_analytics_id": frappe.conf.get("google_analytics_id"),
 			"google_analytics_anonymize_ip": frappe.conf.get("google_analytics_anonymize_ip"),