### Changes

* Refactored & optimized `filter_dynamic_link_doctypes` API
* Optimized query type detection `is_query_type`  - ~40% improvement _(after: 261 ns ± 1.15 ns before: 372 ns ± 4.12 ns)_ 
*  Drop duplicate `get_frontmatter` definition
* Optimized **Login Page**'s get_context - 3x imrpovement _(from 0.047s to 0.017s)_
* Optimized **App Page**'s get_context - ~75% improvement _(0.035s to 0.02s)_
* Refactored `frappe.db.get_singles_dict` - Cast values based on fieldtypes _[accidental bugfix (see usages 😭)]_
* Patch PyPika module once and not on each `frappe.init`
* Refactored `frappe.get_system_settings` - Fetch singles dict and cache in local instead of repeated queries to tabSingles
* New `frappe.get_website_settings`
* Now, add **"Run UI Tests"** label on your PR that updates only Py code to run Cypress tests too - similarly there's **"Run Server Tests"** for the other way 'round.